+++ IronSource/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+ * 7.0.4.1.0
+    * This version of the adapters has been certified with ironSource 7.0.4.1 and MoPub SDK 5.15.0.
+    * Only for side by side mediation cases, ironSource adapters now optionally accept ad unit types to initialize. [Related documentation here.](https://developers.mopub.com/publishers/mediation/networks/ironsource/)
+
  * 7.0.3.1.0
     * This version of the adapters has been certified with ironSource 7.0.3.1 and MoPub SDK 5.14.0.